pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bankai-poc"
version = "0.1.0"
description = "Patch-routed Bankai proof of concept for benchmark-specific reversible XOR patches."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets>=2.21.0",
  "huggingface_hub>=0.24.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
bankai-poc = "bankai_poc.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

